backend/requirements.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Database (PostgreSQL for production)
psycopg2-binary>=2.9,<3.0

# Development tools
python-dotenv>=1.0,<2.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.7,<5.0
